Refactor the Auth pages
