+++ src/components/RegFieldBig/index.js
@@ -2,6 +2,7 @@ import React from 'react';
 import {BlueInput} from './styles';
 
 export default function RegFieldBig(props) {
+
   return (
     <BlueInput
       placeholder={props.placeholder}
