+++ src/pages/Login/index.js
@@ -3,8 +3,10 @@ import Theme, {theme} from '../../../Theme';
 import Background2 from '../../components/Background2';
 import CustomText from '../../components/CustomText';
 import RegFieldBig from '../../components/RegFieldBig';
-import {AuthContext} from '../../contexts/auth'
-import {
+import {AuthContext} from '../../contexts/auth';
+import { TextInput } from 'react-native';
+import { DateContainer } from '../RegistroAluno/styles'
+import{
   LoginContainer,
   ButtonAluno,
   ButtonProfessor,
@@ -42,13 +44,22 @@ export default function Login({navigation}) {
                     value={email}
                     onChangeText={ (text) => setEmail(text) }
               />
-              <RegFieldBig
+              <DateContainer>
+                  <TextInput
                     placeholder="Senha"
                     autoCapitalize="none"
                     value={password}
                     onChangeText={ (text) => setPassword(text) }
                     secureTextEntry={true}
-              />
+                    placeholderTextColor="#F6F6F6"
+                    style={{
+                      color: '#FFFFFF',
+                      fontSize: 14,
+                      flex: 1,
+                      textAlign: 'center',
+                    }}
+                  />
+                </DateContainer>
               <ButtonEntrar onPress={handleLogin}>
                 <CustomText white medium>
                   Entrar
